Allow third-party Java agents under CC
The configuration cache rejected any build whose JVM had a
non-Gradle -javaagent attached, even when the agent merely
registered runtime hooks (idea_rt.jar) or composed with
Gradle's bytecode instrumentation (Jacoco). With Gradle's own
instrumentation agent active, the refusal is unwarranted.
